Clean up mask example script.
