+++ configs/mask_sim.yaml
@@ -1,14 +1,11 @@
-# Grayscale simulation and reconstruction
-
 files:
   original: data/celeba_mini/000019.jpg
   #original: data/original/mnist_3.png
 
-  psf: data/psf/tape_rgb.png
+save: True
   
-
 simulation:
-  object_height: 0.4
+  object_height: 0.25
   # these distance parameters are typically fixed for a given PSF
   scene2mask: 40e-2
   mask2sensor: 4e-3
@@ -24,18 +21,18 @@ simulation:
   # max val in simulated measured (quantized 8 bits)
   max_val: 230
 
-  #wavelength: 523e-9
-  mask_type: "MURA"
-  #size_px: [4056, 3040] # irrelevant for coded aperture
-  size_px: [380, 507] # irrelevant for coded aperture
+  image_format: RGB
+
+  flatcam: True   # only supported if mask.type is "MURA" or "MLS"
+
+
+mask:
+  type: "MURA"    # "MURA", "MLS", "FZA", "PhaseContour"
 
   # Coded Aperture (MURA or MLS)
   #flatcam_method: 'MLS'
   n_bits: 99 # e.g. 8 for MLS, 99 for MURA
-  image_format: RGB
-  flatcam_simulation: True
-  regul: 3e-4
-
+  
   # Phase Contour
   noise_period: [16, 16]
   refractive_index: 1.2
@@ -44,13 +41,19 @@ simulation:
   # Fresnel Zone Aperture
   radius: 0.5  # todo set in m
 
-save: True
 
-admm:
-  n_iter: 20
-  disp_iter: 2
-  # Hyperparameters
-  mu1: 1e-6
-  mu2: 1e-5
-  mu3: 4e-5
-  tau: 0.0001
+recon:
+
+  algo: "tikhonov"   # tikhonov or admm
+
+  tikhonov:
+    reg: 3e-4
+
+  admm:
+    n_iter: 10
+    disp_iter: 2
+    # Hyperparameters
+    mu1: 1e-6
+    mu2: 1e-5
+    mu3: 4e-5
+    tau: 0.0001
